+++ addon/components/comment-form/component.js
@@ -20,6 +20,7 @@ import layout from './template';
  */
 export default Ember.Component.extend({
     layout,
+    i18n: Ember.inject.service(),
     _commentText: null,
 
     actions: {