+++ addon/components/comment-pane/component.js
@@ -30,5 +30,6 @@ import layout from './template';
  */
 export default Ember.Component.extend({
     layout,
+    i18n: Ember.inject.service(),
     comments: null
 });