+++ tests/dummy/app/locales/en-us/translations.js
@@ -1,5 +1,339 @@
 export default {
+    allNodes: {
+        loadMore: "Load more",
+        theseAreTheNodes: "These are the nodes"
+    },
+    collections: {
+        detail: {
+            add: "Add",
+            addProject: "Add project to collection",
+            addRegistration: "Add registration to collection",
+            collectionContents: "Collection Contents",
+            collectionDetails: "Collection Details",
+            created: "created",
+            bookmark: "Bookmark",
+            date: "Date",
+            modifiedLowercase: "modified",
+            modifiedUppercase: "Modified",
+            name: "Name",
+            projectID: "Project id",
+            registrationID: "Registration id",
+            removeProject: "Remove project",
+            removeRegistration: "Remove registration",
+            title: "Title"
+        },
+        index: {
+            collections: "Collections",
+            create: "Create",
+            createCollection: "Create Collection",
+            collectionsList: "Collections list",
+            dateCreated: "Date created",
+            dateModified: "Date modified",
+            title: "Title"
+        }
+    },
+    cookieLogin: {
+        login: "Login",
+        toNodesPage: "To nodes page"
+    },
+    embeddingRecords: {
+        abstract: "Abstract",
+        demonstrationOfEmbedded: "Demonstration of Embedded Records",
+        doi: "DOI",
+        loadANode: "Load a node with an embedded preprint",
+        nodeID: "Node ID",
+        nodesWereEmbedded: "These nodes were embedded in the user response so a separate request was not required",
+        nodeTitle: "Node title",
+        noNodesWarning1: "No nodes with published preprints on them",
+        noNodesWarning2: "that you can access at least",
+        noUsersWarning: "No users matching this query",
+        preprintID: "Preprint ID",
+        preprintWasEmbedded: "This preprint was embedded on this node, so a separate request wasn't required to fetch it",
+        searchResults: "Search results",
+        searchUsers: "Search users and embed nodes",
+        subjects: "Subjects",
+        userNodes: "User Nodes"
+    },
+    institutions: {
+        detail: {
+            authURL: "Auth URL",
+            description: "Description",
+            logoURL: "Logo URL",
+            name: "Name",
+            nodes: "Nodes"
+        },
+        index: {
+            allInstitutions: "All Institutions",
+            authURL: "Auth URL",
+            description: "Description",
+            logoURL: "Logo URL",
+            name: "Name",
+            viewOnOSF: "View on OSF"
+        }
+    },
     login: {
-        pageTitle: 'Login to begin'
+        getFancy: "Get Fancy",
+        loginMessage: "Login",
+        pageTitle: 'Login to begin',
+        signUp: "Sign Up"
+    },
+    nodes: {
+        detail: {
+            children: {
+                components: "Components"
+            },
+            draftRegistrations: {
+                detail: {
+                    attachFile: "Attach File",
+                    beforeYouContinue: "Before you continue",
+                    cancel: "Cancel",
+                    continue: "Continue",
+                    editDraftRegistration: "Edit Draft Registration",
+                    embargoEndDate: "Embargo End Date",
+                    enterEmbargo: "Enter registration into embargo",
+                    initiatedBy: "Initiated by",
+                    lastUpdated: "Last Updated",
+                    makePublic: "Make registration public immediately",
+                    notImplemented: "not implemented",
+                    pleaseChoose: "Please choose",
+                    register: "Register",
+                    registrationChoice: "Registration Choice",
+                    saveAsDraft: "Save as draft",
+                    started: "Started",
+                    youMayAttachMessage1: "You may attach up to 5 files to this question",
+                    youMayAttachMessage2: "You may attach files that you already have in this OSF project, or upload a new file from your computer",
+                    youMayAttachMessage3: "Uploaded files will automatically be added to this project so that they can be registered"
+                },
+                index: {
+                    cancel: "Cancel",
+                    createDraft: "Create Draft",
+                    delete: "Delete",
+                    draftRegistrationsTitle: "Draft Registrations",
+                    edit: "Edit",
+                    initiatedBy: "Initiated by",
+                    lastUpdated: "Last Updated",
+                    newRegistration: "New Registration",
+                    register: "Register",
+                    registrationDesc: "Registration creates a frozen version of the project that can never be edited or deleted but can be withdrawn. Your original project remains editable but will now have the registration linked to it",
+                    selectRegistrationForm: "Continue your registration but selecting a registration form",
+                    started: "Started",
+                    thingsToKnow1: "Ensure your project is in the state you wish to freeze before registering",
+                    thingsToKnow2: "Consider turning links into forks",
+                    thingsToKnow3: "Registrations can have embargo periods for up to four years. If you choose an embargo period, the registration will automatically become public when the embargo expires",
+                    thingsToKnow4: "Withdrawing a registration removes the contents of the registrations but will leave behind a log showing when the registration was created and withdrawn",
+                    thingsToKnowTitle: "Things to know about your registration"
+                }
+            },
+            files: {
+                file: {
+                    checkIn: "Check in",
+                    checkedOutBy: "This file is checked out by a",
+                    checkOut: "Check out",
+                    delete: "Delete",
+                    download: "Download",
+                    tags: "Tags"
+                },
+                provider: {
+                    file: {
+                        revisions: {
+                            download: "Download",
+                            revisionTitle: "Revisions",
+                            size: "Size",
+                            version: "Version"
+                        }
+                    },
+                    index: {
+                        actions: "Actions",
+                        files: "Files",
+                        id: "ID",
+                        path: "Path",
+                        provider: "Provider"
+                    }
+                }
+            },
+            index: {
+                add: "Add",
+                addInstitution: "Add Institution",
+                addLinkedNodeRelationship: "Add linked node relationship",
+                affiliatedInstitutions: "Affiliated Institutions",
+                affiliateNode: "Affiliate Node",
+                analysis: "Analysis",
+                apiDeprecationMessage: "formerly node links. After API v2.1, node links have been deprecated",
+                arrow: "↑",
+                category: "Category",
+                citation: "Citation",
+                closeNodeSettings: "Close Node Settings",
+                communication: "Communication",
+                components: "Components",
+                contrubutors: "Contributors",
+                create: "Create",
+                createComponent: "Create component",
+                data: "Data",
+                dateCreated: "Date Created",
+                dateModified: "Date Modified",
+                deleteProject: "Delete this project",
+                description: "Description",
+                editLicense: "Edit License",
+                editNodeSettings: "Edit node settings",
+                files: "Files",
+                fork: "Fork",
+                forks: "Forks",
+                forkedFrom: "Forked From",
+                forkedOn: "Forked on",
+                hypothesis: "Hypothesis",
+                id: "ID",
+                institutionID: "Institution ID",
+                instrumentation: "Instrumentation",
+                license: "License",
+                linkedNodes: "Linked Nodes",
+                methodsAndMeasures: "Methods and Measures",
+                name: "Name",
+                optional: "optional",
+                other: "Other",
+                preprint: "Preprint",
+                preprints: "Preprints",
+                privacySettings: "Privacy settings",
+                procedure: "Procedure",
+                project: "Project",
+                public: "Public",
+                registrations: "Registrations",
+                remove: "Remove",
+                saveChanges: "Save changes",
+                software: "Software",
+                tags: "Tags",
+                targetNodeID: "Target node ID",
+                title: "Title",
+                unaffiliate: "Unaffiliate",
+                uncategorized: "Uncategorized",
+                viewOnOSF: "View on OSF"
+            },
+            registrations: {
+                noCompletedRegistrations: "There have been no completed registrations of this project. You can start a new registration by clicking the 'New registration' button, and you have the option of saving as a draft registration before submission",
+                pendingEmbargo: "Pending Embargo",
+                pendingRegistration: "Pending Registration",
+                pendingWithdrawal: "Pending Withdrawal",
+                registered: "Registered",
+                registrationsTitle: "Registrations",
+                viewOnOSF: "View on OSF",
+                withdrawn: "Withdrawn"
+            },
+            wiki: {
+                wikis: "Wikis"
+            }
+        },
+        list: {
+            category: "Category",
+            createNode: "Create Node",
+            dateCreated: "Date Created",
+            dateModified: "Date Modified",
+            description: "Description",
+            myNodes: "My Nodes",
+            optional: "optional",
+            title: "Title",
+            viewOnOSF: "View on OSF"
+        }
+    },
+    prereg: {
+        chooseFilesInstead: "Or choose them, if you're not into HTML5",
+        dragFilesHere: "Drag your preregistration files here",
+        name: "Name",
+        selectedFiles: "Selected files",
+        submitPreregistration: "Submit preregistration"
+    },
+    profile: {
+        editProfile: "Edit your profile",
+        editRecord: "Edit record",
+        familyName: "Family Name",
+        fullName: "Full Name",
+        givenName: "Given Name",
+        middleName: "Middle Name",
+        profileInformation: "Profile Information",
+        viewOnOSF: "View on OSF"
+    },
+    registrations: {
+        detail: {
+            category: "Category",
+            dateRegistered: "Date Registered",
+            title: "Title",
+            viewOnOSF: "View on OSF"
+        },
+        index: {
+            category: "Category",
+            dateRegistered: "Date Registered",
+            myRegistrations: "My Registrations",
+            title: "Title",
+            viewOnOSF: "View on OSF"
+        }
+    },
+    userNodes: {
+        user: "User",
+        nodes: "Nodes",
+        loadMore: "Load more"
+    },
+    users: {
+        detail: {
+            affiliatedInstitutions: "Affiliated Institutions",
+            data: "Data",
+            dateRegistered: "Date Registered",
+            fullName: "Full Name",
+            id: "ID",
+            nodes: "Nodes",
+            registrations: "Registrations",
+            title: "Title",
+            viewOnOSF: "View on OSF"
+        },
+        list: {
+            dateRegistered: "Date Registered",
+            fullName: "FullName",
+            users: "Users",
+            viewOnOSF: "View on OSF"
+        }
+    },
+    components: {
+        contribManager: {
+            add: "Add",
+            addContributor: "Add contributor",
+            admin: "Administrator",
+            bibliographic: "Bibliographic",
+            id: "ID",
+            name: "Name",
+            nonBibliographic: "Non-Bibliographic",
+            permissions: "Permissions",
+            read: "Read",
+            readWrite: "Read + Write",
+            remove: "Remove",
+            saveChanges: "Save changes",
+            userID: "User ID"
+        },
+        fileActions: {
+            addFolder: "Add folder",
+            copy: "Copy",
+            delete: "Delete",
+            download: "Download",
+            folderID: "Folder ID",
+            move: "Move",
+            moveTo: "Move to",
+            newFolder: "New folder",
+            newName: "New name",
+            nodeID: "Node ID",
+            optional: "optional",
+            rename: "Rename",
+            replace: "Replace",
+            updateFileContents: "Update file contents",
+            upload: "Upload",
+            uploadFiles: "Upload files"
+        },
+        fileTree: {
+            emptyFolder: "empty folder"
+        },
+        logDetail: {
+            action: "Action",
+            additionalParams: "Additional Params Included",
+            currentNode: "Current Node",
+            date: "Date",
+            originalNode: "Original Node",
+            recentActivity: "Recent Activity",
+            user: "User"
+        }
     }
 };